Got contextual chat to work on a by-url basis
Some restructuring of the data. The main object is a "chat", which exists
on a specific URL (only one per URL, though I don't know how to enforce that).
Each chat also has a thread id associated with it.

A thread contains a bunch of messages.

A message contains a name and the message text.