Added a really crappy way for new users to sign up
Automatically logs them in after they sign up.

Also the name field is now automatically filled based on the signup email.

Need to find a way to assign unique usernames to these users. Firebase is
really not the right tool for this kind of thing. I want to be able to make
constraints about the uniqueness of things, and the existence of values
in other tables. Maybe you can approximate that using the ".validate" rules?